pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "xero-cli"
version = "0.1.0"
description = "AI-powered Xero accounting CLI with Claude integration for transaction classification"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["xero", "accounting", "cli", "ai", "claude", "anthropic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Accounting",
]
dependencies = [
    "typer[all]>=0.12.0",
    "anthropic>=0.25.0",
    "keyring>=25.0.0",
    "platformdirs>=4.0.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "tomli-w>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-httpx>=0.30.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.6.0",
]

[project.scripts]
xero = "xero_cli.main:app"

[project.urls]
Homepage = "https://github.com/dirkbrand/xero-cli"
Repository = "https://github.com/dirkbrand/xero-cli"
Issues = "https://github.com/dirkbrand/xero-cli/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/xero_cli"]

[tool.ruff]
src = ["src"]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
